Phase 1: move function into model widget
